chores: rename task files to that of their functions, add new content to project __init__
