+++ project/__init__.py
@@ -1,9 +1,13 @@
-from .task1 import get_graph_info, create_and_save_two_cycles_graph
-from .task2 import graph_to_nfa, regex_to_dfa
+from .graph_util import get_graph_info, create_and_save_two_cycles_graph
+from .automata_builder import graph_to_nfa, regex_to_dfa
+from .tensor_rpq import AdjacencyMatrixFA, intersect_automata, tensor_based_rpq
 
 __all__ = [
     "get_graph_info",
     "create_and_save_two_cycles_graph",
     "regex_to_dfa",
     "graph_to_nfa",
+    "AdjacencyMatrixFA",
+    "intersect_automata",
+    "tensor_based_rpq",
 ]
