+++ project/automata_builder.py
